Move cmake files to cmake/ dir and move implementations to src/. The latter needs a bit more thought
