+++ cmake/triggeralgsConfig.cmake.in
