+++ cmake/triggeralgsConfigVersion.cmake.in
