+++ src/TriggerActivityMaker_Supernova.cpp
