+++ src/TriggerCandidateMaker_Supernova.cpp
