+++ src/TriggerDecisionMaker_Supernova.cpp
