+++ src/triggeralgs/Supernova/TriggerCandidateMaker_Supernova.hpp
