+++ src/triggeralgs/Supernova/TriggerDecisionMaker_Supernova.hpp
